rename router_adapter.py to abc.py
